% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-threshold_estimation.R
\name{plot.threshold.estimation}
\alias{plot.threshold.estimation}
\title{Plot threshold estimation}
\usage{
\method{plot}{threshold.estimation}(x, model = c("thresholdCS", "thresholdSC", "thresholdSS"), ...)
}
\arguments{
\item{x}{An object of class \code{threshold.estimation}.}

\item{model}{A \code{character} string for the threshold trajectory results to be
displayed. Choose from \code{"thresholdCS"}, \code{"thresholdSC"} or \code{"thresholdSS"}.}

\item{...}{Further arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Displays a visual representation of the threshold estimations obtained by
\code{\link{estimate_threshold}}.
}
\seealso{
\code{\link{estimate_threshold}}
}
