% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-dredge.R
\name{print.best.trajectory}
\alias{print.best.trajectory}
\title{Print best trajectories}
\usage{
\method{print}{best.trajectory}(x, type = c("global", "by_trajectory"), ...)
}
\arguments{
\item{x}{An object of class \code{best.trajectory}.}

\item{type}{A \code{character} string to choose whether to display information about the
final trajectory (\code{"global"}, default) or about the best models of each trajectory
(\code{"by_trajectory"}).}

\item{...}{Further arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Displays a readable overview of the trajectories dredging returned by \code{\link{dredge_trajectories}}
and treated by \code{\link{best_trajectory}}.
}
\seealso{
\code{\link{best_trajectory}}
}
