% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dredge_pglmm.R
\name{print.dredge.table}
\alias{print.dredge.table}
\title{Print dredge table}
\usage{
\method{print}{dredge.table}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dredge.table}.}

\item{...}{Further arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Displays a readable dredge table of \code{\link[phyr]{pglmm}} model obtained by \code{\link{dredge_pglmm}}.
}
\seealso{
\code{\link{dredge_pglmm}}
}
