% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-threshold_estimation.R
\name{print.threshold.estimation}
\alias{print.threshold.estimation}
\title{Print threshold estimation}
\usage{
\method{print}{threshold.estimation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{threshold.estimation}.}

\item{...}{Further arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Displays a readable summary of the threshold estimations obtained by
\code{\link{estimate_threshold}}.
}
\seealso{
\code{\link{estimate_threshold}}
}
