% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senescence.R
\docType{data}
\name{senescence}
\alias{senescence}
\title{Senescence data}
\format{
A \code{data.frame} with 33 rows and 4 columns:
\describe{
  \item{Species}{Species name (capital letters: A, B, C, ..., AG)}
  \item{log_afr}{Simulated logarithm of the age at first reproduction for
  each species}
  \item{log_onset}{Simulated logarithm of the onset of reproductive
  senescence for each species}
  \item{log_mass}{Simulated logarithm of the body mass for each species}
}
}
\source{
Simulated data
}
\usage{
senescence
}
\description{
Senescence is a simulated dataset that provides age at first reproduction
(logarithm), onset of senescence (logarithm) and body mass (logarithm) for
33 species. Species names were replaced with generic labels (A, B, C, etc.)
for illustrative purpose.
}
\keyword{datasets}
