% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-trajectory_formulas.R
\name{trajectory_formulas}
\alias{trajectory_formulas}
\title{Building all trajectory formulas}
\usage{
trajectory_formulas(
  formula,
  trajectory_trait,
  interactions = NULL,
  data,
  trajectories = "all"
)
}
\arguments{
\item{formula}{An object of class \code{formula} representing the constant trajectory formula.}

\item{trajectory_trait}{A character string representing the trajectory trait.
Must exist in \code{data}.}

\item{interactions}{A vector of \code{character} strings representing variables with
which \code{trajectory_trait} interacts. Must exist in \code{data}. Defaults to \code{NULL}.}

\item{data}{A \code{data.frame} object containing the variables named in \code{formula}.}

\item{trajectories}{A \code{character} vector representing the trajectories to be constructed.
Choose among \code{"constant"}, \code{"linear"}, \code{"quadratic"}, \code{"factor"}, \code{"thresholdCS"}, \code{"thresholdSC"}
and \code{"thresholdSS"}. Defaults to all trajectories (\code{"all"}). \code{"thresholdCS"} describes
a "constant-slope" trajectory where the slope is null before the threshold. Similarly,
\code{"thresholdSC"} describes a "slope-constant" trajectory where the slope is null
after the threshold, and \code{"thresholdSS"} authorize a non-null slope both before
and after the threshold.}
}
\value{
An object of class \code{trajectory.formulas} containing:
\describe{
   \item{trajectory_trait}{A \code{character} string representing the trajectory trait.}
   \item{interactions}{The interaction terms.}
   \item{random}{The random terms.}
   \item{data}{The name of the dataset.}
   \item{formulas}{A \code{list} of the trajectory formulas splitted into non-threshold
   and threshold formulas with a \code{trajectory_type} attribute.}
   \item{display}{A readable list of the formulas.}
   }
}
\description{
This function allows to build desired trajectory formulas (constant, linear,
quadratic, factor and threshold) based on a base formula and a defined trait
of interest (the trajectory trait).
}
\examples{
# Provide a base formula, the trait of interest, its interactions, data and the trajectories desired
formulas <- trajectory_formulas(formula = log_onset ~ log_mass + (1|Species),
                                trajectory_trait = "log_afr", interactions = "log_mass",
                                data = senescence,
                                trajectories = c("linear", "quadratic", "thresholdCS")
                                )
}
