% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_ultra.R
\docType{data}
\name{tree_ultra}
\alias{tree_ultra}
\title{Phylogenetic tree (ultrametric): tree_ultra}
\format{
An object of class \code{phylo}, with 33 tips and branch lengths in
millions of years.
}
\source{
TimeTree database (\url{https://www.timetree.org})
}
\usage{
tree_ultra
}
\description{
An ultrametric phylogenetic tree representing evolutionary distances among
33 species. Species names have been anonymised using capital letters (A, B,
C, etc.) for illustrative purpose. This tree can be used for examples
involving phylogenetic models, such as \code{\link[phyr]{pglmm}} as used in \code{\link{dredge_pglmm}}.
}
\details{
The tree was originally generated from the TimeTree database and
then made ultrametric using the \code{\link[phytools]{force.ultrametric}}.
Species names were replaced by generic labels to match those in the \code{\link{senescence}} dataset.
}
\keyword{datasets}
