% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-select_trajectory.R
\name{select_trajectory}
\alias{select_trajectory}
\title{Wrapper for trajectory model selection}
\usage{
select_trajectory(
  formula,
  trajectory_trait,
  interactions = NULL,
  data,
  trajectories = "all",
  start,
  end,
  iter,
  rank = c("AICc", "AIC"),
  delta_criterion = 2,
  adjust_threshold = TRUE,
  lmer_control = NULL
)
}
\arguments{
\item{formula}{An object of class \code{formula} representing the constant trajectory formula.}

\item{trajectory_trait}{A character string representing the trajectory trait.
Must exist in \code{data}.}

\item{interactions}{A vector of \code{character} strings representing variables with
which \code{trajectory_trait} interacts. Must exist in \code{data}. Defaults to \code{NULL}.}

\item{data}{A \code{data.frame} object containing the variables named in \code{formula}.}

\item{trajectories}{A \code{character} vector representing the trajectories to be constructed.
Choose among \code{"constant"}, \code{"linear"}, \code{"quadratic"}, \code{"factor"}, \code{"thresholdCS"}, \code{"thresholdSC"}
and \code{"thresholdSS"}. Defaults to all trajectories (\code{"all"}). \code{"thresholdCS"} describes
a "constant-slope" trajectory where the slope is null before the threshold. Similarly,
\code{"thresholdSC"} describes a "slope-constant" trajectory where the slope is null
after the threshold, and \code{"thresholdSS"} authorize a non-null slope both before
and after the threshold.}

\item{start, end, iter}{\code{Numeric} values controlling the threshold range search:
\itemize{
  \item \code{start}: Lower bound of the threshold range.
  \item \code{end}: Upper bound of the threshold range.
  \item \code{iter}: Increment of the sequence.
  }}

\item{rank}{Criterion used to rank the models. Choose either \code{"AIC"} or \code{"AICc"}.}

\item{delta_criterion}{A \code{numeric} value to define the criterion value within which
the best model is selected.}

\item{adjust_threshold}{Logical. If \code{TRUE} (default), adds 1 parameter (df) and
2 AIC/AICc points to threshold models to account for the implicit cost of the
threshold parameter as suggested by some methods.}

\item{lmer_control}{Advanced users can supply custom optimizer controls for
\code{lmerMod} class objects. Defaults to \code{NULL}, meaning that \code{\link[lme4]{lmer}} uses the
internal object \code{lmerControl_default} whith Bound Optimization BY Quadratic Approximation,
"bobyqa" and 1e5 as maximal number of evaluation.}
}
\value{
An object of class \code{best.trajectory} containing:
\describe{
   \item{best_table}{A \code{data.frame} displaying the rank, number of parameters
   and delta rank value for the best model of each trajectory.}
   \item{final_traj}{A \code{data.frame} displaying the rank, number of parameters
   and delta rank value for the best model across all trajectories and according
   to the \code{delta_criterion} argument.}
   \item{final_traj_name}{A \code{character} string for the name of the final trajectory.}
   \item{summaries}{A \code{list} of \code{summary.merMod} objects (one per trajectory).}
   \item{summary}{A \code{list} of \code{summary.merMod} object for the final trajectory.}
   }
}
\description{
This function allows to build desired trajectory formulas (constant, linear,
quadratic, factor and threshold) based on a base formula and a defined trait
of interest (the trajectory trait), to estimate threshold points for threshold
trajectories, to fit models and to apply \code{\link[MuMIn]{dredge}} on all trajectories,
retaining the best model trajectory.
}
\details{
This function is a wrapper of \code{\link{trajectory_formulas}}, \code{\link{estimate_threshold}},
\code{\link{fit_models}}, \code{\link{dredge_trajectories}} and \code{\link{best_trajectory}}.
Please refer to the help page of the functions above for details.
}
\examples{
res <- select_trajectory(formula = log_onset ~ log_mass + (1|Species),
                         trajectory_trait = "log_afr",
                         interactions = "log_mass",
                         data = senescence,
                         trajectories = "all",
                         start = -0.5, end = 5.7, iter = 0.1,
                         rank = "AICc",
                         delta_criterion = 2,
                         adjust_threshold = TRUE,
                         lmer_control = NULL)

# display an overview of the results
res # type = "global" by default
print(res, type = "by_trajectory")

# display summary (summaries) of retained trajectory (trajectories)
summary(res) # type = "global" by default
summary(res, type = "by_trajectory")
}
